Refactor common code into internal library
