+++ order/main.go
@@ -7,13 +7,11 @@ import (
 	"sort"
 	"time"
 
-	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/scalog/scalog/internal/pkg/kube"
 	"github.com/scalog/scalog/logger"
 	"github.com/scalog/scalog/order/messaging"
 	"github.com/spf13/viper"
@@ -69,15 +67,19 @@ func getRaftIndexPeerUrls() (int, []string) {
 		// FOR TESTING. Single raft node
 		return 1, []string{"http://0.0.0.0:9876"}
 	}
-	clientset := initKubernetesClient()
-	listOptions := metav1.ListOptions{
-		LabelSelector: "app=scalog-order",
+	clientset, err := kube.InitKubernetesClient()
+	if err != nil {
+		logger.Panicf(err.Error())
 	}
-	pods := getShardPods(clientset, listOptions)
+
+	listOptions := metav1.ListOptions{LabelSelector: "app=scalog-order"}
+	pods, err := kube.GetShardPods(clientset, listOptions, viper.GetInt("raft_cluster_size"), viper.GetString("namespace"))
+	if err != nil {
+		logger.Panicf(err.Error())
+	}
+
 	size := len(pods.Items)
-
 	peers := make([]peerIDAndURL, size, size)
-
 	for i, pod := range pods.Items {
 		logger.Printf("Peer ip: " + pod.Status.PodIP + ", uid: " + string(pod.UID))
 		peers[i] = peerIDAndURL{
@@ -106,46 +108,6 @@ func getRaftIndexPeerUrls() (int, []string) {
 	return index, urls
 }
 
-//TODO copy pasted from data/main.go. Move into kubernetes util package
-func initKubernetesClient() *kubernetes.Clientset {
-	// creates the in-cluster config
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		panic(err.Error())
-	}
-	// creates the clientset
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		panic(err.Error())
-	}
-	return clientset
-}
-
-func getShardPods(clientset *kubernetes.Clientset, listOptions metav1.ListOptions) *v1.PodList {
-	for {
-		pods, err := clientset.CoreV1().Pods(viper.GetString("namespace")).List(listOptions)
-		if err != nil {
-			panic(err)
-		}
-		fmt.Println(fmt.Sprintf("looking for %d, but have %d", viper.GetInt("raft_cluster_size"), len(pods.Items)))
-		if len(pods.Items) == viper.GetInt("raft_cluster_size") && allPodsAreRunning(pods.Items) {
-			return pods
-		}
-		//wait for pods to start up
-		time.Sleep(1000 * time.Millisecond)
-	}
-}
-
-func allPodsAreRunning(pods []v1.Pod) bool {
-	for _, pod := range pods {
-		if pod.Status.Phase != v1.PodRunning {
-			return false
-		}
-	}
-	return true
-}
-
-//TODO end copy pasta
 func listenForErrors(errorChannel <-chan error) {
 	for err := range errorChannel {
 		logger.Printf("Raft error: " + err.Error())
